requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
pip==25.1.1
PySocks==1.7.1
pytest==7.4.4